Start using github flavoured markdown.
This also involved moving the models in order to fix a issue where a
node internal was being sent to the browser.
