+++ models/blog.ts
@@ -1,0 +1,12 @@
+export type Post = {
+  postFilename: string;
+  slug: string;
+  isoDate: string;
+  year: string;
+  month: string;
+  day: string;
+  title: string;
+  content: string;
+  categories: string[];
+  tags: string[];
+};
