+++ package.json
@@ -38,6 +38,7 @@
     "react-dom": "^17.0.2",
     "react-helmet": "^6.1.0",
     "remark": "13.0.0",
+    "remark-gfm": "^1.0.0",
     "remark-html": "13.0.1",
     "slugify": "^1.6.0"
   },
