+++ services/markdownToHtml.ts
@@ -1,9 +1,10 @@
 import remark from "remark";
 import html from "remark-html";
+import gfm from "remark-gfm";
 
 export default async function markdownToHtml(
   markdown: string
 ): Promise<string> {
-  const result = await remark().use(html).process(markdown);
+  const result = await remark().use(html).use(gfm).process(markdown);
   return result.toString();
 }
